+++ src/fs-lang/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- better folding (using a foldingProvider instead of declarative regex)
 
 ## [0.1.0] - 2019-07-04
 ### Added